backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
Pillow==11.0.0
mediapipe==0.10.18
opencv-python-headless==4.10.0.84
numpy==1.26.4
httpx==0.28.0
python-dotenv==1.0.1
